Refactor: RpcProfileRepo has typed errors
Also, simulation does not die any more if it tries to create a profile
that already existed

+++ morpheus_storage/src/messages.rs
@@ -26,7 +26,7 @@ impl Envelope {
 }
 
 pub type MessageId = u64;
-pub type ResponseCode = u32;
+pub type ResponseCode = u8;
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct Request<T> {
